Homework(Lecture 7)
