+++ src/Lecture04/DivisibleBy5.java
@@ -1,14 +1,12 @@
-package Homework2;
+package Lecture04;
 
-import java.util.Arrays;
 import java.util.Scanner;
 
-public class LargestNumInArr {
+public class DivisibleBy5 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
 
-        int largestNum = Integer.MIN_VALUE;
-        System.out.println("Please enter the array length:");
+        System.out.println("Please enter the array length (positive number):");
         int length = Integer.parseInt(scanner.nextLine());
 
         int[] numbers = new int[length];
@@ -19,13 +17,15 @@ public class LargestNumInArr {
         }
 
         for (int item : numbers) {
-            if (item > largestNum) {
-                largestNum = item;
+            if (item > 150) {
+                System.out.println("Number greater than 150 encountered and iteration stopped.");
+                break;
+            }
+            if (item % 5 == 0) {
+                System.out.printf("%d is divisible by 5.%n", item);
             }
         }
-        System.out.printf("The largest number in the Array is %d.", largestNum);
     }
+
 }
-
-
