+++ src/Lecture05/DiagonalSum.java
@@ -1,4 +1,4 @@
-package Homework3;
+package Lecture05;
 
 public class DiagonalSum {
     public static void main(String[] args) {
