+++ src/Lecture05/Palindrome.java
@@ -1,4 +1,4 @@
-package Homework3;
+package Lecture05;
 
 import java.util.Scanner;
