+++ src/Lecture06/SmallersNumMethod.java
@@ -1,4 +1,4 @@
-package Homework4;
+package Lecture06;
 
 
 public class SmallersNumMethod {
